requirements.txt
anyio==3.7.1
astor==0.8.1
attrdict==2.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.12.1
bce-python-sdk==0.8.87
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.1.6
cloud-init==23.2.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.0
cryptography==3.4.8
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.0
dbus-python==1.2.18
decorator==5.1.1
distro==1.7.0
distro-info===1.1build1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fire==0.5.0
Flask==2.3.2
flask-babel==3.1.0
fonttools==4.41.1
future==0.18.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
hyperlink==21.0.0
idna==3.3
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.1
numpy==1.25.1
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.5.1
pandas==2.0.3
pdf2docx==0.5.6
pexpect==4.8.0
Pillow==10.0.0
premailer==3.10.0
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pyclipper==1.3.0.post4
pycryptodome==3.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
PyMuPDF==1.20.2
PyMySQL==1.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-docx==0.8.11
python-magic==0.4.24
pytz==2023.3
PyWavelets==1.4.1
PyYAML==5.4.1
rapidfuzz==3.1.2
rarfile==4.0
requests==2.25.1
scikit-image==0.21.0
scipy==1.11.1
SecretStorage==3.3.1
service-identity==18.1.0
shapely==2.0.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sos==4.4
soupsieve==2.4.1
SQLAlchemy==2.0.19
ssh-import-id==5.11
systemd-python==234
termcolor==2.3.0
tifffile==2023.7.18
tqdm==4.65.0
Twisted==22.1.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
visualdl==2.5.3
wadllib==1.3.6
Werkzeug==2.3.6
zipp==1.0.0
zope.interface==5.4.0
